+++ test/Generator/ClassGeneratorTest.php
@@ -12,6 +12,8 @@ use PHPUnit\Framework\TestCase;
 use Reinfi\OpenApiModels\Configuration\Configuration;
 use Reinfi\OpenApiModels\Generator\ClassGenerator;
 use Reinfi\OpenApiModels\Generator\ClassTransformer;
+use Reinfi\OpenApiModels\Generator\NamespaceResolver;
+use Reinfi\OpenApiModels\Generator\OpenApiType;
 
 class ClassGeneratorTest extends TestCase
 {
@@ -23,6 +25,7 @@ class ClassGeneratorTest extends TestCase
     public function testItGeneratesClassesFromOpenApi(): void
     {
         $configuration = new Configuration([], '', '', false);
+        $namespace = new PhpNamespace('Schema');
 
         $openApi = new OpenApi([
             'components' => [
@@ -45,23 +48,36 @@ class ClassGeneratorTest extends TestCase
             $openApi,
             $this->callback(static fn (string $name): bool => in_array($name, ['Test1', 'Test2'], true)),
             $this->isInstanceOf(Schema::class),
-            $this->callback(static fn (PhpNamespace $namespace): bool => $namespace->getName() === 'Schema')
+            $namespace
         );
 
-        $generator = new ClassGenerator($transformer);
+        $namespaceResolver = $this->createMock(NamespaceResolver::class);
+        $namespaceResolver->expects($this->once())->method('initialize')->with($configuration);
+        $namespaceResolver->expects($this->once())->method('resolveNamespace')->with(OpenApiType::Schemas)->willReturn(
+            $namespace
+        );
+
+        $generator = new ClassGenerator($transformer, $namespaceResolver);
 
         $generator->generate($openApi, $configuration);
     }
 
-    public function testItGeneratesClassesFromOpenApiWithNamespace(): void
+    public function testItGeneratesRequestBodies(): void
     {
-        $configuration = new Configuration([], '', 'Api', false);
+        $configuration = new Configuration([], '', '', false);
+        $namespace = new PhpNamespace('RequestBody');
 
         $openApi = new OpenApi([
             'components' => [
-                'schemas' => [
+                'requestBodies' => [
                     'Test1' => [
-                        'type' => 'object',
+                        'content' => [
+                            'application/json' => [
+                                'schema' => [
+                                    'type' => 'object',
+                                ],
+                            ],
+                        ],
                     ],
                 ],
             ],
@@ -72,10 +88,101 @@ class ClassGeneratorTest extends TestCase
             $openApi,
             'Test1',
             $this->isInstanceOf(Schema::class),
-            $this->callback(static fn (PhpNamespace $namespace): bool => $namespace->getName() === 'Api\Schema')
+            $namespace
         );
 
-        $generator = new ClassGenerator($transformer);
+        $namespaceResolver = $this->createMock(NamespaceResolver::class);
+        $namespaceResolver->expects($this->once())->method('initialize')->with($configuration);
+        $namespaceResolver->expects($this->once())->method('resolveNamespace')->with(
+            OpenApiType::RequestBodies
+        )->willReturn($namespace);
+
+        $generator = new ClassGenerator($transformer, $namespaceResolver);
+
+        $generator->generate($openApi, $configuration);
+    }
+
+    public function testItGeneratesResponses(): void
+    {
+        $configuration = new Configuration([], '', '', false);
+        $namespace = new PhpNamespace('Response');
+
+        $openApi = new OpenApi([
+            'components' => [
+                'responses' => [
+                    'Test1' => [
+                        'content' => [
+                            'application/json' => [
+                                'schema' => [
+                                    'type' => 'object',
+                                ],
+                            ],
+                        ],
+                    ],
+                ],
+            ],
+        ]);
+
+        $transformer = $this->createMock(ClassTransformer::class);
+        $transformer->expects($this->once())->method('transform')->with(
+            $openApi,
+            'Test1',
+            $this->isInstanceOf(Schema::class),
+            $namespace
+        );
+
+        $namespaceResolver = $this->createMock(NamespaceResolver::class);
+        $namespaceResolver->expects($this->once())->method('initialize')->with($configuration);
+        $namespaceResolver->expects($this->once())->method('resolveNamespace')->with(
+            OpenApiType::Responses
+        )->willReturn($namespace);
+
+        $generator = new ClassGenerator($transformer, $namespaceResolver);
+
+        $generator->generate($openApi, $configuration);
+    }
+
+    public function testItAppendsSuffixOnMultipleMediaTypes(): void
+    {
+        $configuration = new Configuration([], '', '', false);
+        $namespace = new PhpNamespace('Response');
+
+        $openApi = new OpenApi([
+            'components' => [
+                'responses' => [
+                    'Test1' => [
+                        'content' => [
+                            'application/json' => [
+                                'schema' => [
+                                    'type' => 'object',
+                                ],
+                            ],
+                            'application/xml' => [
+                                'schema' => [
+                                    'type' => 'object',
+                                ],
+                            ],
+                        ],
+                    ],
+                ],
+            ],
+        ]);
+
+        $transformer = $this->createMock(ClassTransformer::class);
+        $transformer->expects($this->exactly(2))->method('transform')->with(
+            $openApi,
+            $this->callback(static fn (string $name): bool => in_array($name, ['Test1Json', 'Test1Xml'], true)),
+            $this->isInstanceOf(Schema::class),
+            $namespace
+        );
+
+        $namespaceResolver = $this->createMock(NamespaceResolver::class);
+        $namespaceResolver->expects($this->once())->method('initialize')->with($configuration);
+        $namespaceResolver->expects($this->once())->method('resolveNamespace')->with(
+            OpenApiType::Responses
+        )->willReturn($namespace);
+
+        $generator = new ClassGenerator($transformer, $namespaceResolver);
 
         $generator->generate($openApi, $configuration);
     }